introduce ember-metal/dictionary
This introduces an inheriting object (typically used with a null
prototype) and hints to v8 (and other runtimes) that it should never go
into “fast” map mode.

At some point, removing this "hint" would be great.

Clearly this last part sounds incorrect, why shouldn’t every go into
fast mode. Unfortunately for unstable objects used as a cache (with a
large number of unknown keys) they may be promoted to fast mode, only to
quickly be demoted again. This is demotion is often costly and unwanted.

Please note this does increase the time it takes to create the
dictionary (over just {})